test: add v2 tests
